MAJOR - routes moved to 'controller'
