+++ .env
@@ -1,5 +1,5 @@
 
 DB_NAME='scheduler_db'
 DB_USER='root'
-DB_PW='password'
+DB_PW='B3n4cc3ss@sql!'
