--- controllers/login.js
@@ -1,44 +1,0 @@
-module.exports = (app, passport) => {
-  app.get('/', (req, res) => {
-    res.render('layouts/main');
-  });
-
-  app.get('/signup', (req, res) => {
-    res.render('signup');
-  });
-
-  app.get('/signin', (req, res) => {
-    res.render('signin');
-  });
-
-  app.post(
-    '/signup',
-    passport.authenticate('local-signup', {
-      successRedirect: '/home',
-      failureRedirect: '/signup'
-    })
-  );
-
-  app.get('/home', isLoggedIn, (req, res) => {
-    res.render('home');
-  });
-
-  app.get('/logout', (req, res) => {
-    req.session.destroy(err => {
-      res.redirect('/');
-    });
-  });
-
-  app.post('/signin',
-    passport.authenticate('local-signin', {
-      successRedirect: '/home',
-      failureRedirect: '/signin'
-    })
-  );
-
-  function isLoggedIn(req, res, next) {
-    if (req.isAuthenticated()) return next();
-
-    res.redirect('/signin');
-  }
-};
