--- routes/index.js
@@ -1,10 +1,0 @@
-const router = require('express').Router();
-const apiRoutes = require('./api');
-
-router.use('/api', apiRoutes);
-
-router.use((req, res) => {
-  res.send("<h1>Wrong Route!</h1><h3>Please try again.</h3>");
-});
-
-module.exports = router;
